admin panel added

+++ app.py
@@ -1,11 +1,13 @@
-from fastapi import FastAPI,Depends
+from fastapi import FastAPI,Depends,HTTPException,Form
 import sqlite3
 from pydantic import BaseModel
 from user_auth import router,decode_verify_token
 from cart import cart_router
 from fastapi.middleware.cors import CORSMiddleware
 
+
 app=FastAPI()
+
 app.add_middleware(
      CORSMiddleware,
      allow_origins=["http://localhost:5173"],
@@ -45,17 +47,26 @@ class Insert_data(BaseModel):
 
 @app.post("/products/")
 def add_product(product_data:Insert_data,current_user:dict=Depends(decode_verify_token)):
-        connection=sqlite3.connect("inventory.db")
-        cursor=connection.cursor()
-        name=product_data.name
-        category=product_data.category
-        quantity=product_data.quantity
-        price=product_data.price
-        new_data=(name,category,quantity,price)
-        cursor.execute("INSERT INTO products(name,category,quantity,price) VALUES(?,?,?,?)",new_data)
-        connection.commit()
-        connection.close()
-        return {"message":"added"}
+        user_role=current_user["userRole"]
+        if user_role=="admin":
+             connection=sqlite3.connect("inventory.db")
+             cursor=connection.cursor()
+             name=product_data.name
+             category=product_data.category
+             quantity=product_data.quantity
+             price=product_data.price
+             new_data=(name,category,quantity,price)
+             cursor.execute("INSERT INTO products(name,category,quantity,price) VALUES(?,?,?,?)",new_data)
+             connection.commit()
+             connection.close()
+             return {"message":"added"}
+        else:
+             raise HTTPException(
+                  status_code=403,
+                  detail="admin access required"
+             )
+                     
+        
 
 @app.get("/view_products")
 def view_products(current_user:dict=Depends(decode_verify_token)):
@@ -72,37 +83,47 @@ def view_products(current_user:dict=Depends(decode_verify_token)):
 
 
 class Update_item(BaseModel):
-     quantity:int
-@app.put("/update_product/{item_id}")
+     quantity:int | None=None
+     price:int | None=None
+@app.patch("/update_product/{item_id}")
 def update_item(product_data:Update_item,item_id:int,current_user:dict=Depends(decode_verify_token)):
-     connection=sqlite3.connect("inventory.db")
-     cursor=connection.cursor()
-     quantity=product_data.quantity
-     data=(quantity,item_id)
-     cursor.execute("UPDATE products SET quantity=? WHERE id=?",(data))
-     connection.commit()
-     cursor.execute("SELECT * FROM products")
-     data=cursor.fetchall()
-     output=[]
-     for i in data:
-          output.append(i)
-     connection.close()
-     return {"message":"updated","Data":output}
+     user_role=current_user["userRole"]
+     if user_role=="admin":
+          connection=sqlite3.connect("inventory.db")
+          cursor=connection.cursor()
+          quantity=product_data.quantity
+          price=product_data.price
+          cursor.execute("UPDATE products SET quantity=? ,price=? WHERE id=?",(quantity,price,item_id))
+          connection.commit()
+          connection.close()
+          return {"message":"updated"}
+     else:
+          raise HTTPException(
+               status_code=403,
+               detail="admin access required"
+          )
 
      
 @app.delete("/delete_product/{item_id}")
 def delete_product(item_id:int,current_user:dict=Depends(decode_verify_token)):
-     connection=sqlite3.connect("inventory.db")
-     cursor=connection.cursor()
-     cursor.execute("DELETE FROM products WHERE id=?",(item_id,))
-     connection.commit()
-     cursor.execute("SELECT * FROM products")
-     data=cursor.fetchall()
-     output=[]
-     for i in data:
-          output.append(i)
-     connection.close()
-     return {"message":"deleted","Data":output}
+     user_role=current_user["userRole"]
+     if user_role=="admin":
+          connection=sqlite3.connect("inventory.db")
+          cursor=connection.cursor()
+          cursor.execute("DELETE FROM products WHERE id=?",(item_id,))
+          connection.commit()
+          cursor.execute("SELECT * FROM products")
+          data=cursor.fetchall()
+          output=[]
+          for i in data:
+               output.append(i)
+               connection.close()
+               return {"message":"deleted","Data":output}
+     else:
+          raise HTTPException(
+               status_code=403,
+               detail="admin access required"
+          )
